split test_base into multiple files
