+++ pandas/tests/indexes/period/test_ops.py
@@ -4,7 +4,7 @@ import pytest
 import pandas as pd
 from pandas import DatetimeIndex, Index, NaT, PeriodIndex, Series
 from pandas.core.arrays import PeriodArray
-from pandas.tests.test_base import Ops
+from pandas.tests.base.test_ops import Ops
 import pandas.util.testing as tm
